% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_rsample.R
\name{ks.rsample}
\alias{ks.rsample}
\alias{ks.rsample.default}
\title{Kolmogorov-Smirnov Tests for Two or More Equally-Sized Samples}
\usage{
ks.rsample(x, ...)

\method{ks.rsample}{default}(x, exact = NULL, simulate.p.value = FALSE, B = 2000, ...)
}
\arguments{
\item{x}{a list of numeric vectors, each containing the data values for one
sample; or a matrix, with each column containing the data values for one sample.
Each sample must be the same size and must not contain NAs.}

\item{...}{further arguments for methods}

\item{exact}{logical; if \code{TRUE}, or if \code{NULL} and \code{(r * D * n) ^ r < 1e4},
then an exact p-value is computed; otherwise, an inexact p-value is computed
according to the setting of \code{simulate.p.value}. (\code{r} is the number of samples,
\code{D} is the test statistic, and \code{n} is the size of each sample)}

\item{simulate.p.value}{logical; ignored in an exact p-value is computed. If
\code{TRUE}, the p-value is estimated using Monte Carlo simulation, and if \code{FALSE},
using the approximation from Böhm and Hornik (2011). \code{FALSE} is recommended
except for small sample sizes, as the approximation is fairly good
at least down to sample
sizes of about 10.}

\item{B}{integer; number of trials for the Monte Carlo p-value estimation}
}
\value{
An object of class "ks.rsample".
}
\description{
Kolmogorov-Smirnov Tests for Two or More Equally-Sized Samples
}
