% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_rsample.R
\name{ks_rsample}
\alias{ks_rsample}
\alias{ks_rsample.default}
\alias{ks_rsample.matrix}
\alias{ks_rsample.formula}
\title{Kolmogorov-Smirnov Tests for Two or More Equally-Sized Samples}
\usage{
ks_rsample(x, ...)

\method{ks_rsample}{default}(x, exact = NULL, simulate.p.value = NULL, B = 10000, ...)

\method{ks_rsample}{matrix}(x, exact = NULL, simulate.p.value = FALSE, B = 10000, ...)

\method{ks_rsample}{formula}(
  formula,
  data,
  exact = NULL,
  simulate.p.value = FALSE,
  B = 10000,
  ...
)
}
\arguments{
\item{x}{a list of numeric vectors, each containing the data values for one
sample; or a matrix, with each column containing the data values for one
sample. Each sample must be the same size, and the data must not include
missing vaues or ties.}

\item{...}{further arguments for methods}

\item{exact}{logical; if \code{TRUE}, or if \code{NULL} and \code{(r * D * n) ^ r < 1e4},
then an exact p-value is computed; otherwise, an inexact p-value is
computed according to the setting of \code{simulate.p.value}. (\code{r} is the number
of samples, \code{D} is the test statistic, and \code{n} is the size of each sample)}

\item{simulate.p.value}{logical; ignored in an exact p-value is computed. If
\code{TRUE}, the p-value is estimated using Monte Carlo simulation, and if
\code{FALSE}, using the approximation from Böhm and Hornik (2011). \code{FALSE} is
recommended except for small sample sizes, as the approximation is fairly
good at least down to sample sizes of about 10.}

\item{B}{integer; number of trials for the Monte Carlo p-value estimation}

\item{formula}{a formula, where the left-hand side is a numeric variable in
\code{data} and the right-hand side contains a single term whose unique values
define the groups for the test. The LHS must not contain any missing
\code{\link[=model.frame]{model.frame()}}.}

\item{data}{an optional data frame (or similar; see \code{\link[=model.frame]{model.frame()}})
containing variables used in \code{formula}.}
}
\value{
An object of class "ks_rsample".
}
\description{
An implementation of Walter Böhm and Kurt Hornik's (2010) two-or-more-sample
Kolmogorov-Smirnov test
}
\details{
For the purpose of calculating the test statistic, \code{model.formula()}
orders the groups by calling \code{\link[=order]{order()}} on the values of the grouping
variables. If the right-hand side of the formula is an interaction term,
each unique level of the interaction is between those variables is taken to
be one group.
}
\examples{
x <- list(rnorm(100, sd = 1), rnorm(100, sd = 2), rnorm(100, sd = 3))
ks_rsample(x)

# using a formula and data frame:
ks_rsample(Sepal.Length ~ Species, data = iris)
}
\references{
Böhm, W., & Hornik, K. (2010). A Kolmogorov-Smirnov Test for r
Samples. \emph{Research Report Series / Department of Statistics and
Mathematics} 105.
\href{https://doi.org/10.57938/2d4c243c-f18c-450e-8d1d-d0a69ff1fe52}{doi:10.57938/2d4c243c-f18c-450e-8d1d-d0a69ff1fe52}
}
